+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- DnsClientGlobalSettings:
+  - Fixed SuffixSearchList Empty String Handling - fixes [Issue #398](https://github.com/PowerShell/NetworkingDsc/issues/398).
 - NetAdapterAdvancedProperty:
   - Removed validation from RegistryKeyword parameter becauase the list
     of valid registry keywords is not fixed and will depend on adapter